WPLUG-125 when dragging, and cut/pasting the teaser container, the existing teaser nodes now no longer disappear (thanks Anden!)

+++ plugins/se.infomaker.ximteaser/TeaserContainerComponent.js
@@ -12,13 +12,21 @@ class TeaserContainerComponent extends Component {
         }
     }
 
+    didMount() {
+        // If teaser container is removed we need to make sure that child nodes also is deleted
+        this.context.editorSession.onRender('document', (change, info, editorSession) => {
+            if (info.action === 'delete' && change.deleted[this.props.node.id]) {
+                editorSession.transaction((tx) => {
+                    this.props.node.nodes.forEach((teaserNodeId) => {
+                        tx.delete(teaserNodeId)
+                    })
+                })
+            }
+        }, this)
+    }
+
     dispose() {
-        // Clean up any remaining Teaser Nodes
-        this.context.editorSession.transaction((tx) => {
-            this.props.node.nodes.forEach((teaserNodeId) => {
-                tx.delete(teaserNodeId)
-            })
-        })
+        this.context.editorSession.off(this)
     }
 
     addTeaser({type}) {
